moves cobra commands into commands package
